refactor: Move block out of utils
- And import type of Block where only type is needed to keep things semantic correct
