+++ frontend/src/components/BlockLayers.vue
@@ -82,9 +82,9 @@
 	</div>
 </template>
 <script setup lang="ts">
+import type Block from "@/block";
 import useBuilderStore from "@/stores/builderStore";
 import useCanvasStore from "@/stores/canvasStore";
-import Block from "@/utils/block";
 import { FeatherIcon } from "frappe-ui";
 import { PropType, ref, watch } from "vue";
 import draggable from "vuedraggable";
