+++ frontend/src/components/MarginHandler.vue
@@ -113,9 +113,9 @@
 	</div>
 </template>
 <script setup lang="ts">
+import type Block from "@/block";
 import { clamp } from "@vueuse/core";
 import { computed, inject, ref, watchEffect } from "vue";
-import Block from "../utils/block";
 import { getNumberFromPx } from "../utils/helpers";
 const props = defineProps({
 	targetBlock: {
